Gave responsiblity of setting dropout to callers
The network had a big problem with dropout. The dropout probability was
given only in the constructor of the class, which means that the
probability remained the same. In practice, if we start the network to
have dropout, we will have it during training and testing.

Now, the caller has to pass the dropout probability to the network. It
is its responsiblity to set the correct probability according to training
or testing stages.

Obs.: a dropout probability of 0.8 means that 80% of the network will
remain. To set dropout off, you have to set the probability to 1.0.